Add cvmix_message_init to set verbosity level
I think it makes the most sense to have cvmix_message only add messages to the
log if the message severity (or verbosity) is larger than a predefined
threshold. The default setting will be not to log anything; if you run
cvmix_message_init() then you will automatically log warnings and errors, and
if you run cvmix_message_init([integer]) you will log messages with a verbosity
greater than or equal to message. The current levels are

1: Verbose (print everything)
2: Diagnostic (print diagnostic values?)
3: EchoNamelist (will be used in cvmix_*_init to echo parameter settings)
4: Warning (print warning messages)
5: Error (print error messages -- up to the GCM for this to trigger an abort)