Added more namelist logging to cvmix_background
If you use Bryan-Lewis, the namelist log will include the 4 parameters used in
BL mixing; if you use a scalar background, static_Mdiff and static_Tdiff will
be logged. No diffusivities are logged for 1D or 2D fields, though the status
of lvary_vertical and lvary_horizontal is logged in all cases.