Modifications to use ParabolicNonLocal option
This commit makes two modifications that improves the Parabolic Non
Local option.  First, the shape function was not computed at layer 1 in
the code and thus led to a very large divergence.  This would create non
physical temperatures in the upper most layer rapidly.  The second
change, which I am less confident in, is to scale the coefficients in
the shape function such that the maximum value is 1.  This is following
the discussion in the CVMix documentation (section 8.6) where
Cs*G(sigma) has a maximum value of 1.  Without this change, the maximum
is about 6.3 (the value of Cs)

Still in testing, only tested in free convection as of now.